language+test.self: Implement binary messages
This commit implements binary messages, which look like this:

foo + bar

And are represented as a MessageNode with a single argument and the
"operator" as the message name.

It is possible to use multiple "operator characters" (as defined in
language/tokens.zig) in a row as a single binary operator:

foo >> 2
foo <$> 1

Currently, the operators are not checked for spaces inbetween, which
should be fixed soon.
Additionally, different operators being mixed without parenthesizing
is also allowed, and will be evaluated right-to-left. This will
eventually be changed, but right now it doesn't really pose a problem.