runtime+language: Initial AST interpreter draft
This is the initial implementation of the AST interpreter. Blocks and
non-primitive methods are still missing.

The interpreter works with reference-counted objects to represent Self
objects. An Object can have a few different types of contents, see
object.zig for details. An Object spawns with 1 ref, gets ref'd by
identifiers and is unref'd by messages.