runtime: Implement blocks & method/block activations + nonlocals
Nonlocals are only partially implemented; they work when their path is
only through blocks and/or methods, but some other stuff which execute
expressions need special handling of it to let it bubble up.

This makes the interpreter more than a little complicated due to the
tracking of method/block activations now, but that was the hardest part
anyway. :^)

Next up is primitives and the standard library.