runtime: Use the actual activation receiver in primitive messages
This should be used in other places as well, since this will inevitably
cause errors in chains of calls. We actually want the receiver to be the
activation object for some messages, like assignment messages, because
the activation object can have mutable slots on it ("local variables" so
to speak). But when a new method activation is done, it should not wrap
an existing method activation object.