runtime: Ref the parent object within activateCommon
In adba734b3dbfe03d43b67da5abcb6ce510a8051, the activation receiver of
the object is taken when it exists and used as the parent in order to
prevent scopes from stacking on top of each other. However this parent
is never properly ref'd; instead the ref was performed by
executeMethodMessage and executeBlockMessage, who don't have any idea
about whether the object has an activation receiver or not. This caused
unmatched refs and derefs resulting in memory leaks.