runtime+utility: Add weak references+weakly ref block's bound method
This commit adds the WeakPtr and WeakPtrBlock types, which allows one to
weakly reference an object. The object does not necessarily need to be
ref-counted; so long as it has a `weak` member which is of type
`WeakPtrBlock` and properly initializes and deinitializes it, it can be
used with `WeakPtr`.

This commit also fixes a reference cycle caused by blocks strongly
holding onto their bound method activation object while being assigned
to a slot of the method activation object itself. A possible case is:

someMethod = (| slot |
    slot: [ "Here is a block which points to the method activation." ].
    "The method activation will now point to the block."
).

The bound method held in the reference is used in figuring out whether
this block's bound method activation still exists or not; if the method
does not exist anymore, the block cannot be executed anyway. Therefore
making the block weakly reference the method is sufficient here.