runtime: Implement proper errors + stack traces
Now we no longer panic whenever an error occurs. Instead, we can raise a
SelfRuntimeError.RuntimeError, which can bubble up to executeScript.
After it has bubbled, the stack trace is printed by looking at the
activation's message name field.

This could be further improved by pointing at the statement which was
executing on each activation when the error occured, but that requires
support from the parser to give us location information.