runtime: Use the bound method as the block activation's receiver
This caused me quite a bit of head-scratching. When the following code
runs on a block:

loop = (value).

whileTrue: blk = (
    [
        value ifFalse: [ ^ nil ].
    ] loop.
)

[ 'foo' ] whileTrue: [ 'bar' ].

`loop` is called on the block declared inside `whileTrue:`. But when
`value` was called within `loop`, the current `self_object` was a method
activation object pointing to that same block declared within
`whileTrue:`. That `self_object` was then incorrectly used as the parent
for the block activation for `value`, causing an infinite loop and a
stack overflow (because the `value` call inside the block within
`whileTrue:` was pointing to the newly-declared block itself).