runtime: Try to find activation receiver for block + allow `value` ident
This commit adds the ability to use `value` within `traits block`
methods to call a block with no arguments. It's an edge case that is
rarely used, but it should be supported regardless. Additionally, the
receiver of a block message is now checked for an activation receiver
first, because it will be wrapped in an activation object when called in
a `traits block` method.