object: Allow objects to be wrapped in multiple activation object layers
Because blocks can be nested, and each block has to access the upper
block's activation object, we now allow objects to be nested in multiple
layers of activation objects. Of course, this should only happen in
blocks; we make sure that the activation receiver is always unpacked
in activateCommon with methods so that a method cannot access another
method's upper scope.