runtime: Clean up remaining object references at exit to make GPA useful
This makes GeneralPurposeAllocator useful again by showing leaks after
all objects have been cleaned up. Of course, this misses the case where
an object itself has been leaked but for now it's good enough. In the
future this won't be needed as we'll be switching to a proper garbage
collector and an object heap.