runtime: Let all the structs handle their own assignable slot copying
This breaks if done in the previous way because the Slots object has no
way of knowing which offset to copy to. Since we are not "subclassing"
Slots (and there is no such thing as classes or virtual methods in Zig),
each object who wants to copy its assignable slots must do that itself.