runtime: Remember references from older spaces to newer spaces
This only happens on objects with assignable slots, which is why I've
only added it there. Assignable slot values will not auto-update because
the object owning the assignable slot is in an older space; we need to
remember references from those objects to newer spaces so that we can
update them when a garbage collection happens.