runtime: Improve handling of newer spaces during older space GC
This is a minor rewrite of the garbage collection code inspired by the
numerous bugs discovered in the heap due to the previously lower quality
of the code here.

The major change here is the addition of a "newer generation link". This
link provides us with the ability to update newer generation spaces'
references and remembered sets when an older generation decides to GC.
This is required because the objects in the older generation might move
and both the remembered set and references from the newer generations to
the older generations will become stale after that move (best case
pointing to forward references, worst case pointing to now-allocated
data).

Other than that, the code has slightly been simplified and code
repetition has been reduced.