runtime: Refactor the interpreter to use completions
The completions idea is borrowed from the ECMA-262 standard. Basically,
the idea is that you have these "completion" objects that interpreter
functions return, and they could signal either that the thing you just
executed did so without anything exceptional, or they could signal that
there's something you need to pay attention to. Then the caller can
simply keep passing it up the stack until it reaches its intended
destination. It's similar to Zig's try mechanism but with extra info
attached.

In this commit I also deleted the superfluous parent activation check in
executeBlockMessage; I realized that if the weak pointer is gone then
that's a perfectly good indication that our parent activation is gone,
so we don't need to do the backwards activation stack walk.