runtime: Always untrack the matching activation object
Just untracking when the execution is successful is insufficient, as
nonlocal returns do not trigger this state. Additionally, untracking the
current self_object is wrong, because the interpreter might have changed
it during any statement. So instead we hold a reference to the exact
tracked value and always untrack it.