runtime: Start stack traces from the origin of the error
Previously, our stack traces would be shifted by 1, causing the actual
location of the error to not be visible. Now that we take source ranges
with each error that is raised, we display that as the topmost
activation's location. The bottommost activation now becomes <top
level>.

Note that this commit removes NonlocalReturn completion types from
execute(Sub)Script handling and turns it into unreachable code instead.