runtime: Add Managed objects + bump ObjectTypeBits
We now have more than 8 object types! :^)

Managed objects allow primitives to ensure that certain values can get a
finalization function called on them when the object is no longer
reachable. It also kind of gives us a way make statically typed things,
but let us hope nobody uses it for nefarious purposes. :^)

The first type of managed value we can hold is a file descriptor. Note
that we can't just do a `close` on the fd and then check for EBADF, that
way lies race conditions (thanks Andrew for being patient with my
initial dismissal :^). So we instead hold whether the fd is already
closed in a flags field next to the actual fd (4 billion fds ought to be
enough for everybody).