runtime: Set limit to arguments in messages + use stack for arg. lists
We now have a limit of 128 arguments in message sends. This wasn't
chosen empirically, it just seemed like a reasonable limit.

More importantly, now that we have a bounded amount, we can have the
argument lists directly in the stack. This allows us to avoid
allocations.

Gives a 14% performance improvement on the fibonacci example, and
greatly reduces system time.