runtime: Use BoundedArray for argument + assignable slot allocations
This gets rid of the short-lived ArgumentValues struct in favor of
std.BoundedArray which is built for things like this. We now enforce a
limit of 255 assignable slots maximum (this was already a limit in the
object code, but wasn't enforced by the interpreter).

One other boon is passing the argument list in getMessageArguments by
pointer rather than having the function return it to us.

With this change and the removal of the remaining allocations in
interpreter code, we get a 14% performance gain on the Fibonacci
example.