runtime: Strip activation object when returning from blocks/methods
When we exit out of a block or method with "self" as the last
expression, we want to send out the receiver. However, while inside a
method or block the self object is the activation object for the current
method/block. Therefore, we should first find the activation receiver,
and then return that instead of the activation object itself.