language+runtime: Implement the _EvaluateString primitive
This is a major step towards getting a REPL up and running.

We can now create Script objects directly from strings in addition to
files. Thankfully the parser and lexer lend themselves pretty well to
this. :^)

Scripts now have a running_path which is used in place of file_path when
resolving subscripts to execute in _RunScript and _EvaluateString. This
is required because the file path no longer represents the actual path
of the script being executed (string scripts get <a string> as their
filename).