language: Only put back part of the symbol slice which was read
It's possible for a 1-byte symbol to be at the very end of the file,
which means we would read less than the representation length. This
would cause us to push back garbage to the PeekStream.

This bug was hidden by the fact that all the Self files in this
repository are terminated with a newline.