runtime+main+objects: Refactor environment into VirtualMachine
This cleans up the slightly-messy environment.zig file we had and turns
it into a proper abstraction. VirtualMachine now carries the allocator,
heap and global objects which are used throughout the virtual machine.

Doing this cleans up a bunch of global state. In addition, I cleaned up
our lookup of traits. Now instead of doing a full lookup on the global
objects for traits then the primitive type, we can directly use the
traits value stored on the VM object.

This gives us a ~15% performance gain on the Fibonacci example.