runtime: Restore stack to previous point on string evaluation error
When a runtime error completion is created, it does not unwind the
stack as it goes up. This is because we want to preserve the stack until
we can print a trace of it. This worked fine in the past where all
errors were fatal; however, now that we are continuing execution after
catching the error in _EvaluateStringIfFail:, this becomes a problem.

We now have an ActivationStack.restoreTo method, which allows us to
restore the stack to a specific point. We do this after we print the
trace of the string evaluation.