language: Attempt to always parse a keyword message after a binary one
Because binary messages bind more strongly than keyword messages we
disallow the parsing of keyword messages for the binary message term.
Afterwards, we should be checking whether it is possible to parse a
keyword message too, so that the keyword message we previously avoided
can now be parsed with the binary message as its receiver.