language: Parse inherited slots and store in AST
This in the first commit in which we start recognizing inherited slots.
They are typed as (syntax subject to change):

    base< = object.

They are not allowed to store method objects (as they expect slots
objects).

Inherited slots are hoisted above non-inherited slots. To facilitate
this, all SlotNode objects now carry an order value.

When an inherited slot is added to an object, the object gains all the
slots of the inherited object; when the inherited object has slots added
to it, every object that inherits from the inherited object gets updated
with the new slots of the inherited object.