runtime: Add support for inherited slots
This builds on the previous few commits to implement inherited slots
in new object creation and in _AddSlots:. _AddSlots: support is not
as thoroughly tested and contains a few broken edge cases.

Note that this version of inherited slots does not currently support
updating existing objects when the inherited object has slots added to
it.