objects: Implement the collection traits+mixins and std vector
This commit implements std traits collection, which is meant to be the
base object which all collections inherit from, similar to the original
Self VM. Unlike the original Self VM however, the base trait does not
assume anything about the mutability, indexability or resizeability of
an object and delegates these extra functionalities to mixins which can
be added to objects alongside std traits collection.

This commit also uses the collection hierarchy to implement std vector,
which is a dynamically growing sequence of values based on std array.