language: Rewrite the lexer into the tokenizer
This commit rewrites the lexer into the Self tokenizer, taking
inspiration from the Zig tokenizer and parser. All the tokenizer does is
figure out the start and end positions of each token; the parser now
handles things like figuring out the values of integers, floating point
numbers, identifiers and strings. This has the advantage of making the
tokenizer completely allocation-free at the cost of making the parser
more complex and having it do a little more work.