runtime: Pass the activation object as receiver to inline methods
We now track the method execution depth (not to be confused with method
activation, as this is the creation of methods). This is used to inform
us on whether a method is inline or not. When a method is
inline (meaning that it's defined a slot of another method), it
expects to receive the activation object of the current scope as its
receiver (so that it can access the methods of its owner slot). This is
more intuitive behavior than before, and several methods in std already
rely on this.