objects: Fix incorrect assumption of parent behavior
This use of private methods relied on the incorrect behavior of the
interpreter where methods would be able to send messages to the
activation objects of upper scopes. Now that this has been fixed, the
way private methods are used in vector cannot work. Instead the messages
must be directly send to the private perspective itself.