runtime: Add the new address' handles to its own space's tracked set
This previously had incorrect behavior because it updated the address
within the handle with the new address without considering whether that
new address was in the same space. updateAllReferencesToImpl now
correctly inserts the handle into the tracked set of the new address'
space so that it can be properly untracked.