runtime: Unify read and assignment lookups
This fixes a problem in which the assignment lookups took "precedence"
over read lookups, and caused unrelated assignments to happen if one
intended to send a message that looks like an assignment which is
earlier in the lookup chain. Now, when we send a message we only perform
one lookup. When we encounter an assignable slot and our message looks
like an assignment, we return that we found an assignment as a lookup
result and the message interpreter handles the rest. Other lookups are
performed as usual. This should result in a small performance increase
as we have effectively halved the lookups.