language: Remove extra newline from source lines
When the parser figures out the line start and end offsets for a
location, it uses the next line's start offset as the "line end".
However this is an error as the start of the next line is the byte after
the newline. Therefore we must decrement the next line start offset by 1
to get the true line end offset.