main+runtime: Implement and switch to a bytecode VM :^)
This completely revamps the interpreter to use bytecode instead of the
AST directly. This allows us to flatten the entire AST into a set of
blocks. This will be very important once the actor model is fully
integrated as it will allow us to pause on any instruction (and an
instruction is fundamental enough to be paused/resumed compared to
individual statements in the AST which can have arbitrary complexity).

This commit also introduces the Actor struct, which replaces
InterpreterContext's role. Once the actor model changes are complete it
will be the basic unit for actors.

Note that this commit has temporarily broken source ranges, and that
will be fixed in future commits.