runtime: Reset the HWM of the latest chunk if it's completely empty
We never move the latest chunk into the unused pool, so its HWM would
never get properly reset if it was completely empty. However, if enough
handles were allocated and freed in the latest chunk such that the chunk
hit its high water mark limit despite having no handles, then we would
unnecessarily allocate a chunk. This commit prevents that by resetting
the high water mark of the latest chunk if free says it has no more
handles and it is full.