runtime: Only advance pc of _EvaluateStringIfFail: sender in error
When we haven't received a runtime error, we return an
ExecutionResult.completion, which automatically advances the pc. However
we also manually advanced the pc, which is only needed for the runtime
error case. This caused us to skip over instructions such as push_arg on
the primitive sender, causing crashes in the REPL.