runtime: Implement values with a known type
This commit adds IntegerValue and ObjectValue types, which allows for
Values with a type that's already known ahead-of-time. This allows us to
skip various safety checks in release mode, while still keeping safety
in debug mode.

This commit also adds various specialized versions of ObjectValue for
easier use throughout the heap.