runtime: Streamline all primitive argument handling :^)
This commit removes a whole bunch of code by getting rid of the
repetitve primitive receiver and argument type checks. A new struct
called PrimitiveArguments, which provides convenience methods for
handling type checking errors, is introduced.

This commit also introduces Object.ObjectT, Object.humanReadableNameFor,
and Object.asType. Object.asType is of particular interest, because it
will allow us to get rid of the repetitive is*Object, as*Object and
mustBe*Object methods below it once everything that uses them in the
runtime is replaced with calls to asType.