runtime: Implement explicit object allocation
This commit implements the full allocation token API as described in
issue #13. The new API works by making all allocations explicit through
a new method on Heap called getAllocation. This method allows a user to
ask for a specific amount of bytes, which the heap will make available
on eden (garbage collecting if necessary) and then return a token object
which gives that piece of code a claim for the requested amount of
bytes.

The AllocationToken object returned by getAllocation allows a user to
allocate through it in either the object or the byte array segment. If
the requested amount of bytes are exceeded, the VM will immediately
crash. This is intended to significantly reduce the amount of problems
caused by implicit garbage collection being allowed anywhere an
allocation happens.

As a side effect, this commit removes a lot of try keywords from the
overall code due to the reduction of fallible API usage (through
heap.allocateInObjectSegment).

Closes #13.