Make the project compatible with both stage1 and stage2 Zig
With these changes the project can now be built on both Zig stage1 and
stage2. :^) The main problems were heap pointer alignment (which stage2
is more correct about), function pointer types and a couple of
outstanding stage2 bugs. Also caught a memory corruption error in the
Zig compiler, which is nice.