runtime: Add globally reachable bit + mark objects globally reachable
This commit makes the top bit of each object's object information word
determine whether the object is globally reachable or not. Globally
reachable objects are intended to be immutable in actor mode.

In order to propagate global reachability, I have implemented a
non-globally reachable object traversal algorithm, which lives in
runtime/object/traversal.zig. This will also be used during object
copying for actor creation and messages once memory protection is
activated there.

Under the following conditions we mark an object graph as globally
reachable:

1. _AddSlots: which adds slots to a globally-reachable slots object.
2. Assigning to an assignable slot of a globally-reachable object.
3. _ArrayAt:Put: which assigns to a globally-reachable array object.

Note that an assignment of a non-globally reachable object to a globally
object will only be allowed to happen in two conditions:

1. Outside of actor mode where anything goes.
2. Inside actor mode, during a global addition process which will stop
   the world (details TBD).

Therefore there shouldn't be a "objects are suddenly untouchable to you"
condition unless the user is intentionally doing it.