runtime: Add a conditional prim. system + disable some prim.s for wasm
Managed to crash the Zig compiler while trying to set this up.

With this change, we can now build a wasm32-wasi version of ZigSelf. :^)
The only remaining challenge is being able to make the main function
read from stdin, and then feeding it a concatenated version of
repl.self (or alternatively, replacing _RunScript with something that
supports a virtual filesystem when we're running under WASI).

Fixes #22.
Fixes #8.