runtime: Refactor bytecode to be more cache-friendly
Lowcode instructions are now stored as a set of individual fields. This
allows for better cache utilization by reducing padding and allowing for
more opcodes/payloads to be stored after each other for better cache
locality.

This only refactors the bytecode side of things, so the interpreter also
need to be refactored to take full advantage of this change.