runtime: Switch the interpreter to a tail-calling model
The tail-calling interpreter should be faster, as the call overhead for
every opcode execution is reduced. Instead of exiting execute everytime
an instruction is completed, we now exit only when an operation that's
not just advancing the instruction index happens.

Additionally, ExecutionResult.success is now removed as it's redundant.

Internally, the system works by generating "specialized executors",
which are execute functions specialized for each opcode. If the
execution is to continue normally, then a tail call is made to the next
instruction's specialized executor, and so on and so forth, until the
interpreter cannot continue.