runtime: Split getSizeForCloning from getSizeInMemory on objects
Previously, we would use the object's own size both for cloning it and
for copying during GC. This is incorrect, because when cloning we want
to copy all the private dependencies of an object as well. For example,
this was causing a byte array's clones to incorrectly share memory.

We now consider cloning and GC copying two different operations and
cloning/bless operations all use the new getSizeForCloning call.