runtime: Refactor interpreter context creation and lifetime
Previously, we would recreate the interpreter context from scratch on
every activation change (yucky!). This also caused a nasty bug which
only triggered in a very rare case (which I plan to add a test case for
soon).

In the meantime, I also took the liberty to simplify how some other core
operations are done in the VM. The caching of the current executable and
block is now done directly within the interpreter context, and we no
longer give up what we're doing to jump back to Actor.executeUntil just
for a refresh. Completion also doesn't have Restart anymore, as that's
been moved to ExecutionResult as well, which centralizes everything that
touches the instruction index to a single place.

Unfortunately seems to cause a 3% performance regression in the
Fibonacci example, which is probably worth investigating.