runtime: Split RuntimeError from completion + use ExecutionResult more
Completion as a concept didn't need to exist anymore thanks to
ExecutionResult, and was just causing confusion. Furthermore, it limited
our ability to specialize our runtime errors based on whether they
needed allocations, because I didn't want to make the Completion
constructors crowded. Now we are able to skip allocations for literal
and comptime-formatted runtime errors.

Most of the code has no functional changes; the opcodes are made to
always return a value now (except the specialized executors, because
Zig's tailcall is still kinda broken).