runtime: Use the slot name bytearray for the method name
Previously we would create the name bytearray twice; once for the slot
name and once for the method name. That was wasteful, so let's just
share them both. We'll need some sort of "VM string" type to prevent
modifications by user code once we have reflection and to optimize
strings further.