runtime: Cache the bytecode block MultiArrayList slices
Previously, we would call MultiArrayList.items() for every single
instruction. Apparently this method is more expensive than a simple
array access, and thus we were doing useless work for every instruction.

This commit introduces the concept of "sealing", which populates the
field caches when the bytecode block is instantiated in field mode. It
ensures in debug mode that no further modifications can be done after
the block has been sealed (the checks are compiled away in release).

This gives about 2.5% of performance improvement with Fibonacci in
release mode, with 5.1% less instructions executed.