Everywhere: Yeet the concept of inherited slots
This was a pretty half-baked idea that doesn't fit well with how Self
works. In particular, there was the issue of what message to send the
inherited slot in order to make it create new "instances" of its data
slots, as well as confusion regarding what to do with multiple inherited
slots and whatnot.

Self is a prototype-based language; it doesn't have a specific
"constructor" and no concept of "instantiation". We will use alternative
programming paradigms to achieve a similar effect to inheritance in
class-based object programming, like data parents.