language+runtime: Pass structs instead of functions around
This switches to a style more in-line with Zig, in light of issue #1717
being closed in Zig (no function expressions). This would be really
complemented with a way to have compile-time interfaces, but we use
anytype for now.

We still have a few exceptions, but they are usually for actually
generating functions at comptime.