runtime: Refactor the object model to reduce sizes
This is a very large refactor that changes most of the object model.

The primary goal of this refactor was to reduce the amount of memory
required by each object, primarily things like maps. This optimizes the
memory requirements of ex-nihilo objects. Additionally, this change
allowed for the removal of the requirement for a map on every object,
which allows for simpler objects to avoid the unnecessary extra machine
word that does nothing.

The major points in this refactor include:
- The most base form of objects is now called BaseObject. Object is a
  kind of BaseObject.
- Maps are no longer Objects, but BaseObjects instead. This allows for
  reducing the amount of bits required to determine the type of a map
  as we no longer have to identify ourselves as a kind of object first.
  Furthermore, since almost all Map kinds are slots-based at the moment,
  the constant and assignable slot count fields are inlined into the map
  header to further reduce sizes.
- ForwardedObject has become a real object for the time being. Once the
  value marker bit refactor is done, this object kind will be removed.
  As part of the requirements for ForwardedObject, all objects are
  currently required to be at least 2 machine words wide; SlotsMap
  currently has an unused machine word for this.
- Since Maps no longer contain a map pointer field on their own and
  objects no longer require a map, the map-map construct is removed.
  Map-map was pretty finnicky to work with and always required special
  cases, and it didn't bring any benefit to the system.

This commit incurs a 2-3% performance penalty on Fibonacci but I haven't
dug into why yet. Performance optimizations will come later.