runtime: Introduce the Float object type
The Float value type never really worked right, and this will allow us
to reduce the value marker size to 1 bit. This will be a bit slower than
integers and we will have to use the object heap, but that's a
compromise we have to take; zigSelf is optimized towards integers.