runtime: Pre-calculate selector hashes during codegen
The SelectorHash construct has been moved to Selector. During AstGen, we
now generate selector hashes for all the selectors we come across. This
way, we only need to compute the hash once for each send site, instead
of every message send.

This gives a significant performance boost in many respects, with a
8-10% speedup on Fibonacci.