runtime: Replace all non-integer sized packed structs with extern struct
Packed structs still behave like integers, in that they can have a
specific alignment and thus padding. In this particular case,
@sizeOf(Slot) returned 32 because it was considered to be a u192 which
is aligned to 32 bytes, which meant that we were allocating memory that
we then didn't use. How this was never caught despite slots being so
ubiquitous in the system is beyond me.

From now on, packed structs are only allowed in code if they contain
their specific backing integer. I'm not sure why the non-sized variant
is allowed in Zig code.