runtime: Rewrite the heap
The new heap is a culmination of my learnings from real-world VMs. The
goal is to apply various techniques and improvements that are present
in production systems, and to improve the code quality.

The heap now contains two generations, instead of three, where the
survivor semi-spaces are part of the new generation. Additionally, byte
arrays are no longer directly allocated on the heap; the heap now
contains only objects.

The old heap had no way to grow its size and panicked when it ran out
of memory. The new heap operates on a set of regions which are allocated
and deallocated as necessary; there's also a recycling mechanism which
attempts to avoid OS heap churn by keeping around regions that have
recently become unused.

Integration tests have been added to make sure the heap behaves
correctly under a number of situations. I intend to add test cases from
future regressions I catch here.

The actual garbage collection code is a bit messy and can be improved;
the intention was to write the most straight-forward code possible to get
things working, and then go back and refactor things where they makes
sense (WET and all that).

Finally, as the focus has been on getting the heap in a working and
correct state, no performance profiling or improvements have been done,
so there's definitely low-hanging fruit to tackle in future commits.