runtime: Add and use unsafe variants of value/object casting methods
Apparently LLVM isn't smart enough to figure that when an optional is
force-unwrapped with .?, any conditional checks that can return `null`
become UB, and thus can be elided. We do LLVM's job here and add
`unsafe` variants for all casting operations, and use them where we
previously did `as*().?`.

5% performance improvement on Fibonacci (but probably everywhere given
Value is ubiquitous).