runtime: Avoid memcpying uninitialized memory when creating MapBuilder
For some reason, RLS breaks down with BoundedArray in particular and
makes Zig memcpy uninitialized memory across stack frames, resulting in
an unnecessary 2KB memcpy every time we instantiate a slots, method or
block object (which is frequent). Didn't actually result in much of a
performance improvement, but it shows less instructions in Callgrind
which is nice.