runtime: Use ArrayList instead of ArrayHashMap for finalization set
As part of the introduction of the new heap, I moved byte arrays out of
the heap, allocating them on the OS heap instead. This has revealed a
big pressure point in the heap which is the finalization set. Since it
was a hash map, and we were potentially allocating thousands of strings,
it meant there was a ton of pressure on it. Additionally, as hash maps
can't be modified in-place, we were forced to re-create the finalization
set in every minor collection.

Then I suddenly had a stroke of inspiration, and realized that we don't
need the finalization set to be a hash map after all. The only times
where we insert new entries is when
`self.markAddressAsNeedingFinalization` is called, which is only in
object constructors. Additionally, we can leverage the fact that the
finalization set is ordered and use a trick to "compact" it in place,
removing items only when we're sure the object's dead. This trick has
been used in both the new and old generations.

Performance results are amazing: 25% improvement in the Fibonacci
example, and 10% better than before the heap rewrite! If we improve upon
how strings are handled, I'm sure we can get tons more improvement here.