runtime: Make trackObject take anytype + validate type at comptime
The previous method of @ptrCast'ing to BaseObject ended up being a
footgun, because Zig allows any pointer to be cast to any other. I ended
up casting a *Value to **BaseObject, which is fully allowed.

Instead, we now accept anytype and do type validation inside trackObject
to make sure we're actually taking something that has BaseObject as its
first machine word, which is way more reliable.