runtime: Allow object and map methods to be omitted
Previously we required all methods to be defined, which meant we had to
stub out some functions which didn't need to be defined for most
objects (namely finalize). Now we have a "default implementation" that
panics. Of course, care must be taken to avoid turning compile-time
errors for required functions into runtime errors, so a followup commit
will introduce interface checking for object and map types.