runtime: Give specialized executors nicer symbol names
If a function is returned from an anonymous struct inside a comptime
function, the name will be something incomprehensible like
"makeSpecializedExecutor__struct_18734". If you return the _struct_
instead, it will actually be given a sensible name containing the
comptime values passed to the function. This makes profiling much
easier (as I can now see how much time is spent where).