runtime: Introduce the value slot concept
This is done in preparation of the inline cache. When a lookup is
performed, we now return a LookupTarget which tells the caller where to
get the actual result from. The user can then call getValueSlot on the
object to get a "ValueSlot" (separate from the physical slots present in
slots objects). The idea is that we cache the ValueSlots obtained as a
result of lookups, and then just return them directly if a cache hit
happens.

This brings us a 20% performance penalty, which is expected (right now
we're just performing an extra dynamic dispatch per send with no gain).
When inline caching is introduced I expect the gains from that to offset
this penalty and more.