runtime: Implement a monomorphic inline cache
This change implements a simple monomorphic inline cache. The cache is
stored on the ExecutableMap, and used in every send instruction. If the
result of a lookup is cacheable, it is put in the inline cache, and the
next time the same send instruction is executed we compare the map of
the receiver against the map we cached; if they are the same then the
lookup result is re-used (by looking up the value slot on the respective
object).

The cache currently causes a non-trivial performance regression due to
various factors:
- Activation objects cannot have caching, causing us to pay an inline
  cache lookup cost for no benefit.
- Block maps are re-created every time the method they reside in is
  activated, completely nullifying inline caches in blocks.

I wanted to push this out anyway in the interest on being able to focus
on other optimizations while still having the necessary groundwork to
improve this (avoiding getting stuck on the same optimization for a long
time). Perhaps I'll put this behind a feature flag so that master
doesn't get a performance degradation.