runtime: Use ObjectDescriptors to create objects
Previously we would use one opcode per slot to push them onto a slot
stack. This meant that we had to interact with the VM state a lot more
to create a single object.

With this change we use a new struct called ObjectDescriptor to create
each object in one go. Because slot values still need arbitrary
evaluation, they are now pushed to the argument stack instead, and
popped in reverse order.

There is still a bunch of useless work we're doing at the interpreter
level because of legacy slot code, which will be cleaned up in follow-up
commits.

No noticeable performance change in the Fibonacci example.